added types and moved Header.tsx
