+++ package.json
@@ -6,6 +6,7 @@
   "dependencies": {
     "@emotion/react": "^11.9.0",
     "@emotion/styled": "^11.8.1",
+    "@material-ui/core": "^4.12.4",
     "@mui/icons-material": "^5.6.2",
     "@mui/material": "^5.6.1",
     "@mui/styled-engine-sc": "^5.5.2",
